#65 decouple more ipc and mpv

+++ src/server/cmd/player/pkg/mpv.go
@@ -21,7 +21,14 @@ func mpvCommand(container MPVSocketFileContainer, trackPath string) (*exec.Cmd, 
 	if !helpers.ProgramExists(mpvProgram) {
 		return nil, errors.New("could not find processor executable on system")
 	}
-	var pars = []string{"--no-video", fmt.Sprintf("--input-ipc-server=%s", container.SocketFileName()), "--quiet", "--pause", trackPath}
+	var pars = []string{
+		"--no-video",
+		fmt.Sprintf("--input-ipc-server=%s", container.SocketFileName()),
+		"--quiet",
+		"--pause",
+		"--keep-open=yes",
+		trackPath,
+	}
 	cmd := exec.Command(mpvProgram, pars...)
 	log.Println("created  processor command with params: ", pars)
 	return cmd, nil
@@ -29,7 +36,17 @@ func mpvCommand(container MPVSocketFileContainer, trackPath string) (*exec.Cmd, 
 
 type MPVProcessor interface {
 	MPVCloser
+	MPVRunner
 	MPVSocketFileContainer
+	MPVStarter
+}
+
+type MPVRunner interface {
+	IsRunning() bool
+}
+
+type MPVStarter interface {
+	Start(filename string) error
 }
 
 type MPVCloser interface {
@@ -41,35 +58,58 @@ type MPVSocketFileContainer interface {
 }
 
 type MPVProcess struct {
-	process    *os.Process
-	socketFile string
+	process   *os.Process
+	isRunning bool
 }
 
-func NewMpvProcessor(filename string) (MPVProcessor, error) {
+func (mpv *MPVProcess) socketfile() string {
+	return filepath.Join(os.TempDir(), "mpv_socket")
+}
+
+func NewMpvProcessor() (MPVProcessor, error) {
 	mpv := &MPVProcess{
-		socketFile: filepath.Join(os.TempDir(), "mpv_socket"),
+		isRunning: false,
+	}
+	return mpv, nil
+}
+
+func (mpv *MPVProcess) Close() error {
+	if mpv.process == nil {
+		return nil
+	}
+	return mpv.process.Kill()
+}
+
+func (mpv *MPVProcess) SocketFileName() string {
+	return mpv.socketfile()
+}
+
+func (mpv *MPVProcess) Start(filename string) error {
+	if mpv.isRunning {
+		if mpv.process != nil {
+			_ = mpv.process.Kill()
+		}
+		mpv.isRunning = false
 	}
 	cmd, err := mpvCommand(mpv, filename)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	if err := cmd.Start(); err != nil {
-		return nil, err
+		return err
 	}
+	defer func() {
+		mpv.isRunning = true
+	}()
 	mpv.process = cmd.Process
 	log.Println("pid: ", mpv.process.Pid)
 
 	// give processor some time to start up
 	sleep(defaultStartupSecs)
 	log.Println("mpv started")
-
-	return mpv, nil
+	return nil
 }
 
-func (mpv *MPVProcess) Close() error {
-	return mpv.process.Kill()
-}
-
-func (mpv *MPVProcess) SocketFileName() string {
-	return mpv.socketFile
+func (mpv *MPVProcess) IsRunning() bool {
+	return mpv.isRunning
 }
